feat: 유저, 모임 정보 수정때 기존 이미지 없으면 에러발생 수정

+++ src/main/java/com/example/moim/club/entity/Club.java
@@ -88,7 +88,9 @@ public class Club extends BaseEntity {
             this.mainEvent = clubUpdateInput.getMainEvent();
         }
         if (profileImgPath != null) {
-            new File(this.profileImgPath).delete();
+            if (this.profileImgPath != null) {
+                new File(this.profileImgPath).delete();
+            }
             this.profileImgPath = profileImgPath;
         }
     }
